Added health bar to both building and units

+++ AdvancedKnights/Advanced_Knights/Assets/Scripts/Buildings.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class Buildings : Selected {
 
     public Player owner;
-    public int maxhealth;
-    public int health;
+    public float maxhealth;
+    public float health;
 
     public GameObject SelectedUI;
